+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Erl2ex is currently pre-alpha software. Expect significant backwards-incompatibl
 
 ## v0.0.10 (not yet released)
 
+*   Compatibility with Elixir 1.3 and Erlang 19.
 *   Support for variables in fun M:F/A.
 *   A macro argument that collides with a reserved word or function name wasn't properly renamed. Fixed.
 *   The "inline" compile option didn't map function names properly. Fixed.